pyproject.toml
[tool.uv.workspace]
members = [
  "stylebook",
]

[tool.uv.sources]
stylebook = { workspace = true }

[dependency-groups]
dev = [
  # lint
  "rulebook-pylint==0.2",
  "pylint==4.0.4",
  "astroid==4.0.2",
  "regex>=2025.11.3",
  "pylint",
  # main
  "PyYAML",
  "pdoc",
  # test
  "pytest",
  "pytest-cov",
]
